+++ app/src/main/java/com/eatssu/android/data/remote/paging/MealReviewPagingSource.kt
@@ -1,54 +1,16 @@
 package com.eatssu.android.data.remote.paging
 
-import androidx.paging.PagingSource
-import androidx.paging.PagingState
-import com.eatssu.android.data.remote.dto.response.toDomain
 import com.eatssu.android.data.remote.service.ReviewService
-import com.eatssu.android.domain.model.Review
-import retrofit2.HttpException
-import java.io.IOException
 
 class MealReviewPagingSource(
-    private val reviewService: ReviewService,
+    reviewService: ReviewService,
     private val mealId: Long?
-) : PagingSource<Int, Review>() {
+) : BaseReviewPagingSource(reviewService) {
 
-    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, Review> {
-        return try {
-            val page = params.key ?: 0
-            val pageSize = params.loadSize
-            
-            val response = reviewService.getMealReviewList(
-                mealId = mealId,
-                page = page,
-                size = pageSize
-            )
-
-            if (response.isSuccess()) {
-                val reviews = response.getOrNull()?.toDomain() ?: emptyList()
-                val hasNext = response.getOrNull()?.hasNext ?: false
-                
-                LoadResult.Page(
-                    data = reviews,
-                    prevKey = if (page == 0) null else page - 1,
-                    nextKey = if (hasNext) page + 1 else null
-                )
-            } else {
-                LoadResult.Error(Exception("Failed to load meal reviews"))
-            }
-        } catch (e: IOException) {
-            LoadResult.Error(e)
-        } catch (e: HttpException) {
-            LoadResult.Error(e)
-        } catch (e: Exception) {
-            LoadResult.Error(e)
-        }
-    }
-
-    override fun getRefreshKey(state: PagingState<Int, Review>): Int? {
-        return state.anchorPosition?.let { anchorPosition ->
-            state.closestPageToPosition(anchorPosition)?.prevKey?.plus(1)
-                ?: state.closestPageToPosition(anchorPosition)?.nextKey?.minus(1)
-        }
-    }
+    override suspend fun executeRequest(page: Int, size: Int) = 
+        reviewService.getMealReviewList(
+            mealId = mealId,
+            page = page,
+            size = size
+        )
 }
